Refactor, fix issues, add page with named orders
