+++ .env.template
@@ -1,0 +1,8 @@
+# server configuration
+OBEDBOT_DEV=1
+OBEDBOT_PORT=4000
+
+# slack config
+OBEDBOT_BOT_TOKEN=
+OBEDBOT_BOT_ID=
+OBEDBOT_CHANNEL_ID=
