+++ package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "@slack/client": "^3.6.0",
     "bluebird": "^3.4.6",
+    "dotenv": "^2.0.0",
     "express": "^4.14.0",
     "html-entities": "^1.2.0",
     "moment": "^2.14.1",
