+++ src/slack.js
@@ -1,46 +1,113 @@
-import {RTM_MESSAGE_SUBTYPES} from '@slack/client';
-import {isNil, find} from 'lodash';
 import database from 'sqlite';
+import moment from 'moment';
+import {isNil, find} from 'lodash';
+import {RTM_MESSAGE_SUBTYPES} from '@slack/client';
 
 import {slack} from './resources';
 import {
-  prettyPrint,
   userExists,
   saveUser,
-  alreadyReacted,
-  stripMention,
   isObedbotMentioned,
+  stripMention,
+  identifyRestaurant,
+  restaurants,
+  prettyPrint,
   isChannelPublic,
-  getTodaysMessages,
+  alreadyReacted,
+  isOrder,
 } from './utils';
-import config from '../config'; // eslint-disable-line import/no-unresolved
+import config from '../config';
 
-/**
- * Checks if the given message is an order
- *
- * @param {string} order - order message
- * @returns {bool} - true if order matches, false if not identified
- */
-export function isOrder(order) {
-  const regexes = config.orderRegex;
-  if (isObedbotMentioned(order)) {
-    order = stripMention(order);
-  }
-  order = order.toLowerCase().trim();
-
-  console.log('Checking order:', order);
-
-  for (let regexKey in regexes) {
-    if (regexes.hasOwnProperty(regexKey)) {
-      if (regexes[regexKey].test(order)) {
-        console.log(`Order type is ${regexKey}`);
-        return true;
+export function loadUsers() {
+  slack.web.channels.info(config.slack.lunchChannelId)
+    .then(async ({channel: {members}}) => {
+      for (let member of members) {
+        if (!(await userExists(member))) {
+          saveUser(member);
+        }
       }
-    }
+    });
+}
+
+export function getTodaysMessages() {
+  let lastNoon = moment();
+  let now = moment();
+
+  // set the date to last Friday if it is Saturday (6), Sunday (0) or Monday (1)
+  if (now.day() === 0 || (now.day() === 1 && now.hours() < 13) || now.day() === 6) {
+    lastNoon.day(lastNoon.day() >= 5 ? 5 : -2);
+  } else if (now.hours() < 13) {
+    lastNoon.subtract(1, 'day');
   }
 
-  console.log('Message is not an order');
-  return false;
+  lastNoon.hours(13);
+  lastNoon.minutes(0);
+  lastNoon.seconds(0);
+
+  const timeRange = {
+    latest: now.valueOf() / 1000,
+    oldest: lastNoon.valueOf() / 1000,
+  };
+
+  return slack.web.channels.history(config.slack.lunchChannelId, timeRange);
+}
+
+export function notifyThatFoodArrived(callRestaurant) {
+  console.log('Notifying that food arrived', callRestaurant);
+  getTodaysMessages()
+    .then(({messages}) => {
+      database.all('SELECT * FROM users')
+        .then((users) => {
+          for (let message of messages) {
+            if (!(isObedbotMentioned(message.text) && isOrder(message.text))) {
+              continue;
+            }
+            const text = stripMention(message.text).toLowerCase().trim();
+            const restaurant = identifyRestaurant(text);
+
+            if (restaurant === callRestaurant ||
+              (callRestaurant === restaurants.presto && restaurant === restaurants.pizza)) {
+              const userChannelId = find(users, ({user_id}) => user_id === message.user).channel_id;
+              if (userChannelId) {
+                slack.web.chat.postMessage(
+                  userChannelId,
+                  `Prišiel ti obed ${text} z ${callRestaurant} :slightly_smiling_face:`,
+                  {as_user: true}
+                );
+              }
+            }
+          }
+        });
+    });
+}
+
+export function notifyThatFoodWontArrive(callRestaurant) {
+  console.log('Notifying that food will not arrive', callRestaurant);
+  getTodaysMessages()
+    .then(({messages}) => {
+      database.all('SELECT * FROM users')
+        .then((users) => {
+          for (let message of messages) {
+            if (!(isObedbotMentioned(message.text) && isOrder(message.text))) {
+              continue;
+            }
+            const text = stripMention(message.text).toLowerCase().trim();
+            const restaurant = identifyRestaurant(text);
+
+            if (restaurant === callRestaurant
+              || (callRestaurant === restaurants.presto && restaurant === restaurants.pizza)) {
+              const userChannelId = find(users, ({user_id}) => user_id === message.user).channel_id;
+              if (userChannelId) {
+                slack.web.chat.postMessage(
+                  userChannelId,
+                  `Dneska bohužiaľ obed z ${callRestaurant} nepríde :disappointed:`,
+                  {as_user: true}
+                );
+              }
+            }
+          }
+        });
+    });
 }
 
 /**
@@ -80,8 +147,9 @@ function removeConfirmation(ts) {
 function privateIsDeprecated(userChannel) {
   slack.web.chat.postMessage(
     userChannel,
-    // eslint-disable-next-line max-len
-    'Objednávanie v súkromných kanáloch bolo vypnuté, pošli prosím svoju objednávku do #obedy :slightly_smiling_face:'
+    'Objednávanie v súkromných kanáloch bolo vypnuté, ' +
+    'pošli prosím svoju objednávku do #obedy :slightly_smiling_face:',
+    {as_user: true}
   );
 }
 
@@ -157,7 +225,7 @@ export async function messageReceived(msg) {
   }
 }
 
-async function processMessages(history) {
+export async function processMessages(history) {
   for (let message of history.messages) {
     console.log(prettyPrint(message));
 
@@ -178,18 +246,12 @@ async function processMessages(history) {
 }
 
 /**
- * Loads the orders since the last noon
- */
-export function loadTodayOrders() {
-  console.log('Loading today\'s orders');
-
-  getTodaysMessages().then(processMessages);
-}
-
-/**
  * Makes the last call for orders
  */
 export function makeLastCall() {
+  if (config.dev) {
+    return;
+  }
   console.log('making last call');
   getTodaysMessages().then(({messages}) => {
     database.all('SELECT * FROM users').then((users) => {
@@ -202,6 +264,7 @@ export function makeLastCall() {
           slack.web.chat.postMessage(
             user.channel_id,
             'Nezabudni si dnes objednať obed :slightly_smiling_face:',
+            {as_user: true}
           );
         }
       }
