+++ views/head.pug
@@ -1,0 +1,10 @@
+head
+  title= title
+  meta(charset='utf-8')
+  link(
+    rel='stylesheet',
+    href='https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css',
+    integrity='sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u',
+    crossorigin='anonymous'
+  )
+  link(rel='stylesheet', href='public/css/style.css')
