+++ views/includes/foodArrived.pug
@@ -1,0 +1,21 @@
+h2(class='title')
+  b Prišlo jedlo!
+div(id='order-buttons' class='row')
+  div
+    a(href="/presto")
+      button(
+        class='btn btn-warning btn-responsive',
+        onclick='return confirm(\'Are you sure?\')'
+      ) Presto
+  div
+    a(href="/veglife")
+      button(
+        class='btn btn-success btn-responsive',
+        onclick='return confirm(\'Are you sure?\')'
+      ) Veglife
+  div
+    a(href="/spaghetti")
+      button(
+        class='btn btn-info btn-responsive',
+        onclick='return confirm(\'Are you sure?\')'
+      ) Spaghetti
