+++ views/includes/foodWontArrive.pug
@@ -1,0 +1,21 @@
+h2(class='title')
+  b Nepríde jedlo :-(
+div(id='order-buttons' class='row')
+  div
+    a(href="/nopresto")
+      button(
+        class='btn btn-warning btn-responsive',
+        onclick='return confirm(\'Are you sure?\')'
+      ) Presto
+  div
+    a(href="/noveglife")
+      button(
+        class='btn btn-success btn-responsive',
+        onclick='return confirm(\'Are you sure?\')'
+      ) Veglife
+  div
+    a(href="/nospaghetti")
+      button(
+        class='btn btn-info btn-responsive',
+        onclick='return confirm(\'Are you sure?\')'
+      ) Spaghetti
