+++ views/includes/links.pug
@@ -1,0 +1,7 @@
+div(class='row')
+  div(class='col-lg-6 col-md-6 col-sm-6 col-xs-6')
+    a(href='/')
+      button(class='btn btn-success btn-responsive')='Home'
+  div(class='col-lg-6 col-md-6 col-sm-6 col-xs-6')
+    a(href='/named')
+      button(class='btn btn-success btn-responsive')='S menami'
