+++ views/restaurants/shop.pug
@@ -1,0 +1,9 @@
+h3(id='title-nakup', class='title')
+  b Nakup
+table(id='table-nakup', class='table table-bordered')
+  tr
+    th Objednávka
+  - for (var order of shop) {
+    tr
+      td= order
+  - }
